add map with f measure

+++ src/otx/algo/detection/heads/atss_head.py
@@ -198,7 +198,7 @@ class ATSSHead(ClassIncrementalMixin, AnchorHead):
             reg_feat = reg_conv(reg_feat)
         cls_score = self.atss_cls(cls_feat)
         # we just follow atss, not apply exp in bbox_pred
-        bbox_pred = scale(self.atss_reg(reg_feat)).float()
+        bbox_pred = scale(self.atss_reg(reg_feat))
         centerness = self.atss_centerness(reg_feat)
         return cls_score, bbox_pred, centerness
 
@@ -233,6 +233,10 @@ class ATSSHead(ClassIncrementalMixin, AnchorHead):
         Returns:
             dict[str, Tensor]: A dictionary of loss components.
         """
+        # cls_scores = [item.float() for item in cls_scores]
+        # bbox_preds = [item.float() for item in bbox_preds]
+        # centernesses = [item.float() for item in centernesses]
+
         featmap_sizes = [featmap.size()[-2:] for featmap in cls_scores]
         if len(featmap_sizes) != self.prior_generator.num_levels:
             msg = "featmap_sizes and self.prior_generator.num_levels have different levels."
